+++ ntex/tests/web_httpserver.rs
@@ -1,25 +1,16 @@
-use net2::TcpBuilder;
 use std::sync::mpsc;
-use std::{net, thread, time::Duration};
+use std::{thread, time::Duration};
 
 #[cfg(feature = "openssl")]
 use open_ssl::ssl::SslAcceptorBuilder;
 
+use ntex::server::TestServer;
 use ntex::web::{self, App, HttpResponse, HttpServer};
-
-fn unused_addr() -> net::SocketAddr {
-    let addr: net::SocketAddr = "127.0.0.1:0".parse().unwrap();
-    let socket = TcpBuilder::new_v4().unwrap();
-    socket.bind(&addr).unwrap();
-    socket.reuse_address(true).unwrap();
-    let tcp = socket.to_tcp_listener().unwrap();
-    tcp.local_addr().unwrap()
-}
 
 #[cfg(unix)]
 #[ntex::test]
 async fn test_start() {
-    let addr = unused_addr();
+    let addr = TestServer::unused_addr();
     let (tx, rx) = mpsc::channel();
 
     thread::spawn(move || {
@@ -92,7 +83,7 @@ fn ssl_acceptor() -> std::io::Result<SslAcceptorBuilder> {
 async fn test_start_ssl() {
     use ntex::web::HttpRequest;
 
-    let addr = unused_addr();
+    let addr = TestServer::unused_addr();
     let (tx, rx) = mpsc::channel();
 
     thread::spawn(move || {