+++ ntex/CHANGES.md
@@ -4,6 +4,8 @@
 
 * HTTP1 dispatcher refactoring
 
+* Replace net2 with socket2 crate
+
 ## [0.1.1] - 2020-04-01
 
 * Project fork